[Refactor] java doc 추가
