+++ src/test/java/com/nhnacademy/bookpubshop/member/dummy/MemberDummy.java
@@ -2,18 +2,21 @@ package com.nhnacademy.bookpubshop.member.dummy;
 
 import com.nhnacademy.bookpubshop.tier.dummy.TierDummy;
 import com.nhnacademy.bookpubshop.member.entity.Member;
+import com.nhnacademy.bookpubshop.tier.entity.Tier;
 import java.time.LocalDateTime;
 
 /**
- * Some description here
+ * 회원 더미 클래스
+ * 테스트시 해당 클래스를 쉽게 생성하기 위한 클래스 입니다.
+ * 동료들과 협력하여 코딩할 때 유용합니다.
  *
  * @author : 임태원
  * @since : 1.0
  **/
 public class MemberDummy {
-    public static Member dummy() {
+    public static Member dummy(Tier tier) {
         return new Member(null,
-                TierDummy.dummy(),
+                tier,
                 "id",
                 "nickname",
                 "taewon",
